11 JUN - Eliminación del Home_v2.jsx, Catalog_v3.jsx y Contact_v2.jsx
