+++ frontend/src/components/Footer/Footer.css
@@ -212,26 +212,35 @@
 }
 
 /* CONTACT INFO */
-.contact-info {
+.footer-image {
     display: flex;
-    flex-direction: column;
-    gap: 0.75rem;
-}
-
-.contact-item {
-    display: flex;
+    justify-content: center;
     align-items: flex-start;
-    gap: 0.75rem;
-    font-size: 0.875rem;
-    line-height: 1.5
 }
 
-.contact-icon {
-    width: 1.25rem;
-    height: 1.25rem;
-    color: var(--primary-color);
-    flex-shrink: 0;
-    margin-top: 0.125rem;
+.footer-img {
+    width: 100%;
+    max-width: 300px;
+    height: 100%;
+    object-fit: cover;
+    border-radius: 1rem;
+    box-shadow: 0 4px 20px rgba(0, 0, 0, 0.3);
+    transition: transform 0.3s;
+}
+
+.footer-img:hover {
+    transform: scale(1.02);
+}
+
+@media (max-width: 768px) {
+    .footer-image {
+        order: -1;
+    }
+
+    .footer-img {
+        max-width: 250px;
+        height: 150px;
+    }
 }
 
 /* FOOTER BOTTOM */
