--- frontend/src/pages/Catalog_v3.css
@@ -1,218 +1,0 @@
-.catalog-v2-page {
-    padding: 2rem 0;
-    background-color: #f8f9fa;
-    min-height: 100vh;
-}
-
-.catalog-header {
-    text-align: center;
-    margin-bottom: 3rem;
-}
-
-.catalog-title {
-    font-size: 2.5rem;
-    font-weight: 700;
-    color: var(--heading-color);
-    margin-bottom: 0.5rem;
-}
-
-.catalog-subtitle {
-    font-size: 1.125rem;
-    color: var(--text-color);
-}
-
-/* FILTER SECTION */
-.filter-section {
-    margin-bottom: 2rem;
-}
-
-.filter-buttons {
-    display: flex;
-    gap: 0.75rem;
-    flex-wrap: wrap;
-    justify-content: center;
-}
-
-.filter-btn {
-    padding: 0.75rem 1.5rem;
-    border: 2px solid #e5e7eb;
-    background: white;
-    color: var(--text-color);
-    border-radius: 2rem;
-    font-weight: 500;
-    cursor: pointer;
-    transition: all 0.3s;
-}
-
-.filter-btn:hover {
-    border-color: var(--primary-color);
-    color: var(--primary-color);
-}
-
-.filter-btn:active {
-    background: var(--primary-color);
-    border-color: var(--primary-color);
-    color: white;
-}
-
-/* PRODUCTS GRID */
-.products-grid-v2 {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(280px, 1fr));
-    gap: 1.5rem;
-}
-
-.product-card-v2 {
-    background: white;
-    border-radius: 1rem;
-    overflow: hidden;
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-    transition: transform 0.3s, box-shadow 0.3s;
-    position: relative;
-}
-
-.product-card-v2:hover {
-    transform: translateY(-4px);
-    box-shadow: 0 8px 24px rgba(0, 0, 0, 0.15);
-}
-
-/* isNew */
-/*
-.new-badge {
-    position: absolute;
-    top: 1rem;
-    left: 1rem;
-    background: var(--yellow-color);
-    color: white;
-    padding: 0.25rem 0.75rem;
-    border-radius: 1rem;
-    font-size: 0.75rem;
-    font-weight: 600;
-    z-index: 2;
-}
-*/
-
-.product-image-v2 {
-    height: 250px;
-    position: relative;
-    overflow: hidden;
-    background: #f8f9fa;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.product-image-v2 img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    transition: transform 0.3s;
-}
-
-.product-card-v2:hover .product-image-v2 img {
-    transform: scale(1.05);
-}
-
-.product-overlay {
-    position: absolute;
-    top: 1rem;
-    right: 1rem;
-    opacity: 0;
-    transition: opacity 0.3s;
-}
-
-.product-card-v2:hover .product-overlay {
-    opacity: 1;
-}
-
-.wishlist-btn {
-    background: white;
-    border: none;
-    border-radius: 50%;
-    width: 40px;
-    height: 40px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    cursor: pointer;
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-    transition: all 0.3s;
-}
-
-.wishlist-btn:hover {
-    background: var(--primary-color);
-    color: white;
-}
-
-.product-info-v2 {
-    padding: 1.25rem;
-}
-
-.product-brand {
-    font-size: 0.875rem;
-    color: var(--primary-color);
-    font-weight: 600;
-    margin-bottom: 0.5rem;
-}
-
-.product-name-v2 {
-    font-size: 1.125rem;
-    font-weight: 600;
-    color: var(--heading-color);
-    margin-bottom: 0.75rem;
-    line-height: 1.4;
-}
-
-.product-rating {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    margin-bottom: 0.75rem;
-}
-
-.stars {
-    display: flex;
-    gap: 0.125rem;
-}
-
-.star-filled {
-    color: var(--yellow-color);
-}
-
-.star-empty {
-    color: #e5e7eb;
-}
-
-.rating-text {
-    font-size: 0.875rem;
-    color: var(--text-color);
-}
-
-.product-price-v2 {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    margin-bottom: 0.5rem;
-}
-
-.current-price {
-    font-size: 1.25rem;
-    font-weight: 700;
-    color: var(--heading-color);
-}
-
-.add-to-cart-btn {
-    width: 100%;
-    background: var(--primary-color);
-    color: white;
-    border: none;
-    padding: 0.75rem;
-    border-radius: 0.5rem;
-    font-weight: 600;
-    cursor: pointer;
-    transition: background-color 0.3s;
-}
-
-.add-to-cart-btn:hover {
-    background: #0056d6;
-}
