+++ frontend/src/pages/Contact.css
@@ -1,59 +1,190 @@
 .contact-page {
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-    background-color: #f9fafb;
+    background: #f8f9fa;
 }
 
-.contact-main {
-    flex-grow: 1;
-    padding: 2rem 0;
+/* HERO SECTION */
+.contact-hero {
+    background: linear-gradient(135deg, var(--primary-color) 0%, #0056d6 100%);
+    color: white;
+    padding: 4rem 0;
+    text-align: center;
 }
 
-/*
-.container {
-    max-width: 1440px;
+.hero-content {
+    max-width: 600px;
     margin: 0 auto;
-    padding: 0 1rem;
-}
-*/
-
-.contact-card {
-    background-color: white;
-    border-radius: 0.5rem;
-    overflow: hidden;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
 }
 
-.contact-card-inner {
-    display: flex;
-    flex-direction: column;
+.hero-title {
+    font-size: 3rem;
+    font-weight: 700;
+    margin-bottom: 1rem;
 }
 
-@media (min-width: 768px) {
-    .contact-card-inner {
-        flex-direction: row;
+.hero-description {
+    font-size: 1.125rem;
+    opacity: 0.9;
+    line-height: 1.6;
+}
+
+@media (max-width: 768px) {
+    .hero-title {
+        font-size: 2rem;
     }
 }
 
-.contact-info {
-    padding: 1.5rem;
+/* CONTACT METHODS */
+.contact-methods {
+    padding: 4rem 0;
+    background: white;
+}
+
+.methods-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+    gap: 2rem;
+}
+
+.method-card {
+    text-align: center;
+    padding: 2rem;
+    border-radius: 1rem;
+    background: #f8f9fa;
+    transition: all 0.3s;
+    border: 2px solid transparent;
+}
+
+.method-card:hover {
+    transform: translateY(-4px);
+    box-shadow: 0 8px 32px rgba(0, 0, 0, 0.1);
+    border-color: var(--primary-color);
+}
+
+.method-icon-wrapper {
+    width: 80px;
+    height: 80px;
+    background: linear-gradient(135deg, var(--primary-color), #0056d6);
+    border-radius: 50%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    margin: 0 auto 1.5rem;
+}
+
+.method-icon {
+    width: 36px;
+    height: 36px;
+    color: white;
+}
+
+.method-title {
+    font-size: 1.25rem;
+    font-weight: 600;
+    color: var(--heading-color);
+    margin-bottom: 0.5rem;
+}
+
+.method-description {
+    color: var(--text-color);
+    margin-bottom: 1rem;
+    font-size: 0.875rem;
+}
+
+.method-info {
+    font-weight: 600;
+    color: var(--heading-color);
+}
+
+.method-info.link {
+    color: var(--primary-color);
+    text-decoration: none;
+    transition: color 0.3s;
+}
+
+.method-info.link:hover {
+    color: #0056d6;
+}
+
+/* MAIN CONTENT */
+.contact-main {
+    padding: 4rem 0;
+}
+
+.contact-content {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    gap: 4rem;
+    align-items: start;
+}
+
+@media (max-width: 1024px) {
+    .contact-content {
+        grid-template-columns: 1fr;
+        gap: 3rem;
+    }
+}
+
+/* FORM SECTION */
+.contact-form-section {
+    background: white;
+    border-radius: 1rem;
+    overflow: hidden;
+    box-shadow: 0 4px 20px rgba(0, 0, 0, 0.1);
+}
+
+.section-tabs {
+    display: flex;
+    border-bottom: 1px solid #e5e7eb;
+}
+
+.tab-btn {
+    flex: 1;
+    padding: 1rem 1.5rem;
+    border: none;
+    background: #f8f9fa;
+    color: var(--text-color);
+    font-weight: 500;
+    cursor: pointer;
+    transition: all 0.3s;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 0.5rem;
+}
+
+.tab-btn.active {
+    background: white;
+    color: var(--primary-color);
+    border-bottom: 2px solid var(--primary-color);
+}
+
+.tab-btn:hover:not(.active) {
+    background: #e9ecef;
+}
+
+.form-container,
+.faq-container {
+    padding: 2rem;
+}
+
+.form-title,
+.faq-title {
+    font-size: 1.5rem;
+    font-weight: 700;
+    color: var(--heading-color);
+    margin-bottom: 0.5rem;
+}
+
+.form-subtitle {
+    color: var(--text-color);
+    margin-bottom: 2rem;
+    line-height: 1.6;
+}
+
+.contact-form {
     display: flex;
     flex-direction: column;
     gap: 1.5rem;
-}
-
-@media (min-width: 768px) {
-    .contact-info {
-        width: 50%;
-    }
-}
-
-.contact-title {
-    font-size: 1.5rem;
-    font-weight: 600;
-    color: #1f2937;
-    margin-bottom: 1rem;
 }
 
 /* 2x2 GRID LAYOUT */
@@ -171,11 +302,53 @@
     height: 1.25rem;
 }
 
-.contact-map {
-    height: 16rem;
-    width: 100%;
-    border-radius: 0.5rem;
+/* FAQ SECTION */
+.faq-list {
+    display: flex;
+    flex-direction: column;
+    gap: 1.5rem;
+}
+
+.faq-item {
+    padding: 1.5rem;
+    background: #f8f9fa;
+    border-radius: 0.75rem;
+    border-left: 4px solid var(--primary-color);
+}
+
+.faq-question {
+    display: flex;
+    align-items: center;
+    gap: 0.75rem;
+    font-size: 1rem;
+    font-weight: 600;
+    color: var(--heading-color);
+    margin-bottom: 0.75rem;
+}
+
+.faq-icon {
+    width: 1.25rem;
+    height: 1.25rem;
+    color: var(--primary-color);
+}
+
+.faq-answer {
+    color: var(--text-color);
+    line-height: 1.6;
+    margin: 0;
+}
+
+/* MAP SECTION */
+.map-section {
+    background: white;
+    border-radius: 1rem;
     overflow: hidden;
+    box-shadow: 0 4px 20px rgba(0, 0, 0, 0.1);
+}
+
+.map-container {
+    height: 300px;
+    position: relative;
 }
 
 .map-iframe {
@@ -184,20 +357,19 @@
     border: 0;
 }
 
-.contact-form-container {
-    background-color: var(--primary-color);
-    padding: 1.5rem;
-    color: white;
+.map-info {
+    padding: 2rem;
 }
 
-@media (min-width: 768px) {
-    .contact-form-container {
-        width: 50%;
-    }
-}
-
-.form-title {
-    font-size: 1.5rem;
+.map-title {
+    font-size: 1.125rem;
     font-weight: 600;
-    margin-bottom: 1.5rem;
+    color: var(--heading-color);
+    margin-bottom: 0.75rem;
+}
+
+.map-description {
+    color: var(--text-color);
+    line-height: 1.6;
+    margin-bottom: 1.5rem
 }
