--- frontend/src/pages/Contact_v2.css
@@ -1,355 +1,0 @@
-.contact-v2-page {
-    background: #f8f9fa;
-}
-
-/* HERO SECTION */
-.contact-hero {
-    background: linear-gradient(135deg, var(--primary-color) 0%, #0056d6 100%);
-    color: white;
-    padding: 4rem 0;
-    text-align: center;
-}
-
-.hero-content {
-    max-width: 600px;
-    margin: 0 auto;
-}
-
-.hero-title {
-    font-size: 3rem;
-    font-weight: 700;
-    margin-bottom: 1rem;
-}
-
-.hero-description {
-    font-size: 1.125rem;
-    opacity: 0.9;
-    line-height: 1.6;
-}
-
-@media (max-width: 768px) {
-    .hero-title {
-        font-size: 2rem;
-    }
-}
-
-/* CONTACT METHODS */
-.contact-methods {
-    padding: 4rem 0;
-    background: white;
-}
-
-.methods-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-    gap: 2rem;
-}
-
-.method-card {
-    text-align: center;
-    padding: 2rem;
-    border-radius: 1rem;
-    background: #f8f9fa;
-    transition: all 0.3s;
-    border: 2px solid transparent;
-}
-
-.method-card:hover {
-    transform: translateY(-4px);
-    box-shadow: 0 8px 32px rgba(0, 0, 0, 0.1);
-    border-color: var(--primary-color);
-}
-
-.method-icon-wrapper {
-    width: 80px;
-    height: 80px;
-    background: linear-gradient(135deg, var(--primary-color), #0056d6);
-    border-radius: 50%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin: 0 auto 1.5rem;
-}
-
-.method-icon {
-    width: 40px;
-    height: 40px;
-    color: white;
-}
-
-.method-title {
-    font-size: 1.25rem;
-    font-weight: 600;
-    color: var(--heading-color);
-    margin-bottom: 0.5rem;
-}
-
-.method-description {
-    color: var(--text-color);
-    margin-bottom: 1rem;
-    font-size: 0.875rem;
-}
-
-.method-info {
-    font-weight: 600;
-    color: var(--heading-color);
-}
-
-.method-info.link {
-    color: var(--primary-color);
-    text-decoration: none;
-    transition: color 0.3s;
-}
-
-.method-info.link:hover {
-    color: #0056d6;
-}
-
-/* MAIN CONTENT */
-.contact-main {
-    padding: 4rem 0;
-}
-
-.contact-content {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 4rem;
-    align-items: start;
-}
-
-@media (max-width: 1024px) {
-    .contact-content {
-        grid-template-columns: 1fr;
-        gap: 3rem;
-    }
-}
-
-/* FORM SECTION */
-.contact-form-section {
-    background: white;
-    border-radius: 1rem;
-    overflow: hidden;
-    box-shadow: 0 4px 20px rgba(0, 0, 0, 0.1);
-}
-
-.section-tabs {
-    display: flex;
-    border-bottom: 1px solid #e5e7eb;
-}
-
-.tab-btn {
-    flex: 1;
-    padding: 1rem 1.5rem;
-    border: none;
-    background: #f8f9fa;
-    color: var(--text-color);
-    font-weight: 500;
-    cursor: pointer;
-    transition: all 0.3s;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 0.5rem;
-}
-
-.tab-btn.active {
-    background: white;
-    color: var(--primary-color);
-    border-bottom: 2px solid var(--primary-color);
-}
-
-.tab-btn:hover:not(.active) {
-    background: #e9ecef;
-}
-
-.form-container,
-.faq-container {
-    padding: 2rem;
-}
-
-.form-title,
-.faq-title {
-    font-size: 1.5rem;
-    font-weight: 700;
-    color: var(--heading-color);
-    margin-bottom: 0.5rem;
-}
-
-.form-subtitle {
-    color: var(--text-color);
-    margin-bottom: 2rem;
-    line-height: 1.6;
-}
-
-.contact-form-v2 {
-    display: flex;
-    flex-direction: column;
-    gap: 1.5rem;
-}
-
-.form-row {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 1rem;
-}
-
-@media (max-width: 768px) {
-    .form-row {
-        grid-template-columns: 1fr;
-    }
-}
-
-.form-group {
-    display: flex;
-    flex-direction: column;
-}
-
-.form-label {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    font-weight: 500;
-    color: var(--heading-color);
-    margin-bottom: 0.5rem;
-    font-size: 0.875rem;
-}
-
-.label-icon {
-    width: 1rem;
-    height: 1rem;
-    color: var(--primary-color);
-}
-
-.form-input,
-.form-textarea {
-    padding: 0.75rem 1rem;
-    border: 2px solid #e5e7eb;
-    border-radius: 0.5rem;
-    font-size: 0.875rem;
-    transition: all 0.3s;
-    background: white;
-}
-
-.form-input:focus,
-.form-textarea:focus {
-    outline: none;
-    border-color: var(--primary-color);
-    box-shadow: 0 0 0 3px rgba(0, 103, 255, 0.1);
-}
-
-.form-textarea {
-    resize: vertical;
-    min-height: 120px;
-}
-
-.submit-btn {
-    background: linear-gradient(135deg, var(--primary-color), #0056d6);
-    color: white;
-    border: none;
-    padding: 1rem 2rem;
-    border-radius: 0.5rem;
-    font-weight: 600;
-    cursor: pointer;
-    transition: all 0.3s;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 0.5rem;
-    font-size: 1rem;
-}
-
-.submit-btn:hover {
-    transform: translateY(-2px);
-    box-shadow: 0 8px 24px rgba(0, 103, 255, 0.3);
-}
-
-/* FAQ SECTION */
-.faq-list {
-    display: flex;
-    flex-direction: column;
-    gap: 1.5rem;
-}
-
-.faq-item {
-    padding: 1.5rem;
-    background: #f8f9fa;
-    border-radius: 0.75rem;
-    border-left: 4px solid var(--primary-color);
-}
-
-.faq-question {
-    display: flex;
-    align-items: center;
-    gap: 0.75rem;
-    font-size: 1rem;
-    font-weight: 600;
-    color: var(--heading-color);
-    margin-bottom: 0.75rem;
-}
-
-.faq-icon {
-    width: 1.25rem;
-    height: 1.25rem;
-    color: var(--primary-color);
-}
-
-.faq-answer {
-    color: var(--text-color);
-    line-height: 1.6;
-    margin: 0;
-}
-
-/* MAP SECTION */
-.map-section {
-    background: white;
-    border-radius: 1rem;
-    overflow: hidden;
-    box-shadow: 0 4px 20px rgba(0, 0, 0, 0.1);
-}
-
-.map-container {
-    height: 300px;
-    position: relative;
-}
-
-.map-iframe {
-    width: 100%;
-    height: 100%;
-    border: 0;
-}
-
-.map-info {
-    padding: 2rem;
-}
-
-.map-title {
-    font-size: 1.125rem;
-    font-weight: 600;
-    color: var(--heading-color);
-    margin-bottom: 0.75rem;
-}
-
-.map-description {
-    color: var(--text-color);
-    line-height: 1.6;
-    margin-bottom: 1.5rem
-}
-
-.location-details {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-}
-
-.location-item {
-    display: flex;
-    align-items: center;
-    gap: 0.75rem;
-    font-size: 0.875rem;
-}
-
-.location-icon {
-    width: 1.25rem;
-    height: 1.25rem;
-    color: var(--primary-color);
-}
