+++ frontend/src/pages/Home.css
@@ -1,104 +1,354 @@
-.hero__content {
+.home {
+    overflow-x: hidden;
+}
+
+/* HERO SECTION */
+.hero-section {
+    background: linear-gradient(135deg, #f8f9fa 0%, #e9ecef 100%);
+    padding: 4rem 0;
+    min-height: 90vh;
     display: flex;
-    flex-direction: column;
-    gap: 90px;
+    align-items: center;
+}
+
+.hero-content {
+    display: flex;
     align-items: center;
     justify-content: space-between;
-    padding-top: 60px;
+    gap: 4rem;
 }
 
-.hero__text {
-    max-width: 100%;
+@media (max-width: 1024px) {
+    .hero-content {
+        flex-direction: column;
+        text-align: center;
+        gap: 3rem;
+    }
 }
 
-.hero__title {
-    font-size: 36px;
-    line-height: 46px;
-    color: var(--heading-color);
+.hero-text {
+    flex: 1;
+    max-width: 600px;
+}
+
+.hero-title {
+    font-size: 3.5rem;
     font-weight: 800;
+    color: var(--heading-color);
+    line-height: 1.2;
+    margin-bottom: 1.5rem;
 }
 
-.hero__counter {
-    margin-top: 30px;
+.highligth {
+    color: var(--primary-color);
+    position: relative;
+}
+
+.highligth::after {
+    content: '';
+    position: absolute;
+    bottom: 0;
+    left: 0;
+    width: 100%;
+    height: 3px;
+    color: var(--primary-color);
+    border-radius: 2px;
+}
+
+.hero-description {
+    font-size: 1.125rem;
+    line-height: 1.7;
+    color: var(--text-color);
+    margin-bottom: 2rem;
+}
+
+.hero-buttons {
     display: flex;
-    flex-direction: column;
-    gap: 20px;
+    gap: 1rem;
+    margin-bottom: 3rem;
+    justify-content: center;
 }
 
-.counter__item h2 {
-    font-size: 36px;
-    line-height: 56px;
+@media (max-width: 768px) {
+    .hero-buttons {
+        flex-direction: column;
+        align-items: center;
+    }
+}
+
+.btn {
+    padding: 1rem 2rem;
+    border-radius: 50px;
+    font-weight: 600;
+    text-decoration: none;
+    border: none;
+    cursor: pointer;
+    transition: all 0.3s;
+    font-size: 1rem;
+    background: var(--primary-color);
+    color: white;
+}
+
+.btn:hover {
+    background: #0056d6;
+    transform: translateY(-2px);
+}
+
+.btn-large {
+    padding: 1.25rem 2.5rem;
+    font-size: 1.125 rem;
+}
+
+/* HERO STATS */
+.hero-stats {
+    display: flex;
+    gap: 2rem;
+}
+
+@media (max-width: 768px) {
+    .hero-stats {
+        justify-content: center;
+        flex-wrap: wrap;
+    }
+}
+
+.stat-item {
+    text-align: left;
+}
+
+@media (max-width: 1024px) {
+    .stat-item {
+        align-items: center;
+    }
+}
+
+.stat-number {
+    font-size: 2.5rem;
     font-weight: 700;
     color: var(--heading-color);
+    margin-bottom: 0.5rem;
 }
 
-.counter__bar {
+.stat-bar {
     width: 100px;
     height: 8px;
-    border-radius: 9999px;
-    display: block;
-    margin-top: -14px;
+    border-radius: 4px;
+    margin-bottom: 0.5rem;
 }
 
-.counter__bar--yellow {
-    background-color: var(--yellow-color);
+.stat-bar.yellow {
+    background: var(--yellow-color);
 }
 
-.counter__bar--purple {
-    background-color: var(--purple-color);
+.stat-bar.purple {
+    background: var(--purple-color);
 }
 
-.counter__bar--iris {
-    background-color: var(--iris-color);
+.stat-bar.iris {
+    background: var(--iris-color);
 }
 
-.hero__images {
+.stat-label {
+    color: var(--text-color);
+    font-size: 0.875rem;
+}
+
+/* HERO IMAGES */
+.hero-images {
     display: flex;
-    gap: 30px;
-    justify-content: flex-end;
+    gap: 2rem;
+    flex: 1;
+    max-width: 500;
 }
 
-.hero__image-col {
-    margin-top: 30px;
+.image-column {
+    display: flex;
+    flex-direction: column;
+    gap: 1.5rem;
 }
 
-.hero__image {
+.hero-img {
+    border-radius: 1rem;
+    box-shadow: 0 8px 32px rgba(0, 0, 0, 0.1);
+    transition: transform 0.3s;
+}
+
+.hero-img:hover {
+    transform: scale(1.02);
+}
+
+.hero-img.main {
     width: 100%;
+    height: 400px;
+    object-fit: cover;
 }
 
-.hero__image-col img:first-child {
-    margin-bottom: 30px;
+.hero-img.secondary {
+    width: 100%;
+    height: 180px;
+    object-fit: cover;
 }
 
-@media only screen and (min-width: 1024px) {
-    .hero__content {
-        flex-direction: row;
-    }
-
-    .hero__text {
-        max-width: 570px;
-    }
-
-    .hero__title {
-        font-size: 60px;
-        line-height: 70px;
-    }
-
-    .hero__counter {
-        margin-top: 70px;
-        flex-direction: row;
-        align-items: center;
-        gap: 30px;
-    }
-
-    .counter__item h2 {
-        font-size: 44px;
-        line-height: 54px;
-    }
+/* FEATURES SECTION */
+.features-section {
+    padding: 5rem 0;
+    background: white;
 }
 
-@media only screen and (min-width: 1536px) {
-    .hero__section {
-        height: 800px;
+.section-header {
+    text-align: center;
+    margin-bottom: 3rem;
+}
+
+.section-title {
+    font-size: 2.5rem;
+    font-weight: 700;
+    color: var(--heading-color);
+    margin-bottom: 1rem;
+}
+
+.section-subtitle {
+    font-size: 1.125rem;
+    color: var(--text-color);
+    max-width: 600px;
+    margin: 0 auto;
+}
+
+.features-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fix, minmax(300px, 1fr));
+    gap: 2rem;
+}
+
+.feature-card {
+    text-align: center;
+    padding: 2rem;
+    border-radius: 1rem;
+    background: #f8f9fa;
+    transition: transform 0.3s, box-shadow 0.3s;
+}
+
+.feature-card:hover {
+    transform: translateY(-4px);
+    box-shadow: 0 8px 32px rgba(0, 0, 0, 0.1);
+}
+
+.feature-icon-wrapper {
+    width: 80px;
+    height: 80px;
+    background: var(--primary-color);
+    border-radius: 50%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    margin: 0 auto 1.5rem;
+}
+
+.feature-icon {
+    width: 36px;
+    height: 36px;
+    color: white;
+}
+
+.feature-title {
+    font-size: 1.25rem;
+    font-weight: 600;
+    color: var(--heading-color);
+    margin-bottom: 1rem;
+}
+
+.feature-description {
+    color: var(--text-color);
+    line-height: 1.6;
+}
+
+/* BRANDS SECTION */
+.brands-section {
+    padding: 5rem 0;
+    background: #f8f9fa;
+}
+
+.brands-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+    gap: 2rem;
+}
+
+.brand-card {
+    background: white;
+    padding: 2rem;
+    border-radius: 1rem;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+    transition: transform 0.3s;
+}
+
+.brand-card:hover {
+    transform: translateY(-2px);
+}
+
+.brand-card img {
+    max-width: 100%;
+    max-height: 100%;
+    object-fit: contain;
+    filter: grayscale(100%);
+    transition: filter 0.3s;
+}
+
+.brand-card:hover img {
+    filter: grayscale(0%);
+}
+
+/* CTA SECTION */
+.cta-section {
+    padding: 5rem 0;
+    background: linear-gradient(135deg, var(--primary-color) 0%, #0056d6 100%);
+    color: white;
+}
+
+.cta-content {
+    text-align: center;
+    max-width: 600px;
+    margin: 0 auto;
+}
+
+.cta-title {
+    font-size: 2.5rem;
+    font-weight: 700;
+    margin-bottom: 1rem;
+}
+
+.cta-description {
+    font-size: 1.125rem;
+    margin-bottom: 2rem;
+    opacity: 0.9;
+}
+
+/* RESPONSIVE */
+@media (max-width: 768px) {
+    .hero-title {
+        font-size: 2.5rem;
+    }
+
+    .section-title {
+        font-size: 2rem;
+    }
+
+    .cta-title {
+        font-size: 2rem;
+    }
+
+    .hero-images {
+        max-width: 100%;
+    }
+
+    .hero-img.main {
+        height: 300px;
+    }
+
+    .hero-img.secondary {
+        height: 140px;
     }
 }
